property-mapper-mvp/backend/app/data_import/requirements.txt
pandas
openpyxl
python-dotenv
SQLAlchemy
psycopg
asyncpg